Move unit tests to live alongside application code
This makes it easier to understand what has test coverage vs what does
not
