+++ lib/build.test.js
@@ -1,5 +1,5 @@
-const { expect, log, fixturePath } = require('./test');
-const { Builder } = require('../lib/build');
+const { expect, log, fixturePath } = require('../test/test');
+const { Builder } = require('./build');
 
 const sinon = require('sinon');
