+++ lib/platform.test.js
@@ -1,6 +1,6 @@
 const sinon = require('sinon');
-const { expect } = require('./test');
-const { PLATFORMS, Platform, platformForName } = require('../lib/platform');
+const { expect } = require('../test/test');
+const { PLATFORMS, Platform, platformForName } = require('./platform');
 
 describe('platform.js', () => {
 	beforeEach(async () => {
