+++ lib/runner.test.js
@@ -1,7 +1,7 @@
-const { expect, config, log, fixturePath } = require('./test');
+const { expect, config, log, fixturePath } = require('../test/test');
 const proxyquire = require('proxyquire');
 const { Runner } = proxyquire('../lib/runner', { './config': { config } });
-const { RunMode } = require('../lib/config');
+const { RunMode } = require('./config');
 
 const sinon = require('sinon');
